tweaks

+++ civitai-cache/cache.py
@@ -13,6 +13,7 @@ MONGODB_CONNECTION = os.getenv('MONGODB_CONNECTION',"mongodb://localhost:27017")
 MONGODB_DATABASE = os.getenv('MONGODB_DATABASE',"database")
 MODELS_COLLECTION = os.getenv('MODELS_COLLECTION',"models")
 CHECK_HASH = os.getenv('CHECK_HASH',False)
+IGNORE_LOCKS = os.getenv('IGNORE_LOCKS',False)
 
 client = MongoClient(MONGODB_CONNECTION)
 db = client[MONGODB_DATABASE]
@@ -50,43 +51,70 @@ for model in models:
                         with open(f"{fileDirectory}/file.json", "wb") as f:
                             f.write(json_util.dumps(file).encode('utf-8'))
                             fileName = f"{fileDirectory}/{file['name']}"
+                            hashFileName = f"{fileDirectory}/{file['name']}.sha256"
                             download = False
                             if os.path.isfile(fileName):
                                 print(f"File {fileName} already exists.")
                                 if CHECK_HASH:
-                                    print(f"Checking SHA256 hash...")
-                                    h = sha256(fileName).lower()
-                                    s = file['hashes']['SHA256'].lower()
-                                    if s != h:
-                                        print(f"🛑 Hash {s} expected but got {h}.  Will redownload...")
-                                        download = True
+                                    if os.path.isfile(hashFileName):
+                                        print(f"Hash for {fileName} already checked.")
                                     else:
-                                        print("✅ Hash good.")
+                                        print(f"Checking SHA256 hash...")
+                                        h = sha256(fileName).lower()
+                                        s = file['hashes']['SHA256'].lower()
+                                        if s != h:
+                                            print(f"🛑 Hash {s} expected but got {h}.  Will redownload...")
+                                            download = True
+                                        else:
+                                            print(f"✅ Hash good.  Saving to {hashFileName}")
+                                            with open(hashFileName, "w") as file:
+                                                # Write the hash to the file
+                                                file.write(h)
                             else:
                                 download = True
                             
                             if download:
-                                print(f"Downloading {fileName}")
-                                # Send a request to the URL with streaming enabled
-                                response = requests.get(file["downloadUrl"], stream=True)
-                                # Get the total file size from the response headers
-                                total_size = int(response.headers.get('content-length', 0))
-                                # Define a chunk size (in bytes) for each iteration
-                                chunk_size = 1024
-                                # Create a progress bar
-                                progress = tqdm(total=total_size, unit='B', unit_scale=True)
-                                # Write the response content to a file in chunks
-                                with open(fileName, 'wb') as f:
-                                    for chunk in response.iter_content(chunk_size=chunk_size):
-                                        # Write the chunk to the file
-                                        f.write(chunk)
-                                        
-                                        # Update the progress bar
-                                        progress.update(len(chunk))
+                                lockFileName = f"{fileDirectory}/{file['name']}.lock"
+                                if not os.path.exists(lockFileName):
+                                    print(f"Downloading {fileName}")
 
-                                # Close the progress bar
-                                progress.close()
-                                print(f'{fileName} has been downloaded successfully.')
+                                    # Send a request to the URL with streaming enabled
+                                    with open(lockFileName, "w") as lockfile:
+                                        # Write the hash to the file
+                                        lockfile.write("")
+
+                                    response = requests.get(file["downloadUrl"], stream=True)
+                                    # Get the total file size from the response headers
+                                    total_size = int(response.headers.get('content-length', 0))
+                                    # Define a chunk size (in bytes) for each iteration
+                                    chunk_size = 1024
+                                    # Create a progress bar
+                                    progress = tqdm(total=total_size, unit='B', unit_scale=True)
+                                    # Write the response content to a file in chunks
+                                    with open(fileName, 'wb') as f:
+                                        for chunk in response.iter_content(chunk_size=chunk_size):
+                                            # Write the chunk to the file
+                                            f.write(chunk)
+                                            
+                                            # Update the progress bar
+                                            progress.update(len(chunk))
+
+                                    # Close the progress bar
+                                    progress.close()
+                                    print(f'{fileName} has been downloaded successfully.')
+                                    if os.path.exists(lockFileName):
+                                        os.remove(lockFileName)
+                                    h = sha256(fileName).lower()
+                                    s = file['hashes']['SHA256'].lower()
+                                    if s != h:
+                                        print(f"🛑 Hash {s} expected but got {h}.  Still...")
+                                    else:
+                                        print(f"✅ Hash good.  Saving to {hashFileName}")
+                                        with open(hashFileName, "w") as file:
+                                            # Write the hash to the file
+                                            file.write(h)
+                                else:
+                                    print(f"{lockFileName} detected.  Skipping download.")
     except:
         tb = traceback.format_exc()
         print(f"Could not save {model['id']}\n{tb}")
